% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{computePathwayScores}
\alias{computePathwayScores}
\title{Title}
\usage{
computePathwayScores(pageRanks, pathwayMembership, distanceFunction,
  distanceFunctionExtraArgs)
}
\arguments{
\item{distanceFunctionExtraArgs}{}
}
\description{
Title
}

