% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{computePersonalizationVectors}
\alias{computePersonalizationVectors}
\title{Title}
\usage{
computePersonalizationVectors(experimentalData, PPIGraph, defaults,
  scoreFunction, scoreFunctionExtraArgs)
}
\arguments{
\item{scoreFunctionExtraArgs}{}
}
\description{
Title
}

