% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{downloadReactomeInteractionsMITAB}
\alias{downloadReactomeInteractionsMITAB}
\title{Title}
\usage{
downloadReactomeInteractionsMITAB(data_folder, species)
}
\arguments{
\item{species}{}
}
\description{
Title
}

