% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{getGeneInteractionListFromReactomeMITAB}
\alias{getGeneInteractionListFromReactomeMITAB}
\title{Title}
\usage{
getGeneInteractionListFromReactomeMITAB(mitab, usmap)
}
\arguments{
\item{usmap}{}
}
\description{
Title
}

