% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{getGeneInteractionsFromIRefMITAB}
\alias{getGeneInteractionsFromIRefMITAB}
\title{Get a gene interaction data frame from a iRef ppi mitab data frame}
\usage{
getGeneInteractionsFromIRefMITAB(mitab)
}
\arguments{
\item{mitab}{@param mitab - The mitab data frame containing PPI's e.g
Obtained using
\url{http://www.inside-r.org/packages/cran/iRefR/docs/get_irefindex}}
}
\value{
a data frame that contains the gene interaction
  graph in two columns, the first being the source HGNCSymbol and the second
  being the target HGNCSymbol of an interaction
}
\description{
Get a gene interaction data frame from a iRef ppi mitab data frame
}
\examples{
\dontrun{
library(iRefR)
iref_mitab=get_irefindex(tax_id="9606",data_folder="/tmp/",iref_version = "13.0")
irefGeneInteractions = getGeneInteractionListFromIRefMITAB(iref_mitab)
}
}

