% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{getGeneInteractionsFromReactomeMITAB}
\alias{getGeneInteractionsFromReactomeMITAB}
\title{Get a gene interaction data frame from a reactome ppi mitab data frame}
\usage{
getGeneInteractionsFromReactomeMITAB(mitab, usmap)
}
\arguments{
\item{mitab}{- The mitab data frame containing PPI's
See  \code{\link{downloadReactomeInteractionsMITAB}}}

\item{usmap}{- A mapping file between UniProt ID's and HGNC Symbols
See \code{\link{getUniProtToHGNCSymbolMapping}}}
}
\value{
a data frame that contains the gene interaction
  graph in two columns, the first being the source HGNCSymbol and the second
  being the target HGNCSymbol of an interaction
}
\description{
Get a gene interaction data frame from a reactome ppi mitab data frame
}
\examples{
\dontrun{
     data_folder=tempdir()
     usmap<-getUniProtToHGNCSymbolMapping(data_folder)
     mitab <- downloadReactomeInteractionsMITAB(data_folder)
     geneInteractions <- getGeneInteractionsFromReactomeMITAB(mitab,usmap)
}
}
\seealso{
\code{\link{downloadReactomeInteractionsMITAB}}
}

