% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{importExpressionDataGEO}
\alias{importExpressionDataGEO}
\title{Import expression data and probe mappings from GEO dataset}
\usage{
importExpressionDataGEO(eset, gpl, selectorFunction = probeSelectorATSAT,
  combinerFunction = probeCombinerMean)
}
\arguments{
\item{eset}{An expression set, for more details refer to
\url{https://www.bioconductor.org/packages/3.3/bioc/vignettes/Biobase/inst/doc/ExpressionSetIntroduction.pdf}
 and \url{http://svitsrv25.epfl.ch/R-doc/library/GEOquery/html/GDS2MA.html}}

\item{gpl}{GEO Platform entity, refer
\url{http://svitsrv25.epfl.ch/R-doc/library/GEOquery/html/GPL-class.html}
for more details}

\item{selectorFunction}{a function that accepts a probeName string and
returns true if the probe is to be considered. See
\code{\link{probeSelectorATSAT}} for an example.}

\item{combinerFunction}{a function that accepts a comma separated string and
returns true if the probe is to be considered. See
\code{\link{probeCombinerMean}} for an example.}
}
\value{
A data frame with columns as sample names, and rows
  containing the expression values of each as HGNC Symbol.
}
\description{
This function processes the expression set and gpl files obtained using the
GEOQuery package
}
\examples{
\dontrun{
library(GEOquery)
data_folder = tempdir()
gds = getGEO('GDS3837',AnnotGPL = TRUE, getGPL = TRUE,destdir = data_folder)
gpl = getGEO(Meta(gds)$platform,destdir = data_folder)
eset = GDS2eSet(gds, GPL=gpl,do.log2=FALSE)
gdata=importExpressionDataGEO(eset,gpl)
}
}

