% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{loadExperimentData}
\alias{loadExperimentData}
\title{Load experiment data}
\usage{
loadExperimentData(expdata, hgnc2probe, selectorFunction = probeSelectorATSAT,
  combinerFunction = probeCombinerMean)
}
\arguments{
\item{expdata}{a probeName x sample matrix of expression values
expression data}

\item{hgnc2probe}{a data frame with two columns probeName and HGNCSymbol
for mapping probeNames to HGNC Symbols}

\item{selectorFunction}{a function that accepts a probeName string and
returns true if the probe is to be considered. See
\code{\link{probeSelectorATSAT}} for an example.}

\item{combinerFunction}{a function that accepts a comma separated string and
returns true if the probe is to be considered. See
\code{\link{probeCombinerMean}} for an example.}
}
\value{
A data frame with columns as sample names, and rows as HGNC Symbols
 containing the expression values for analysis
}
\description{
Load experiment data
}

