% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{outdegreeNormalizedFCOneMinusPScoreFunction}
\alias{outdegreeNormalizedFCOneMinusPScoreFunction}
\title{Title}
\usage{
outdegreeNormalizedFCOneMinusPScoreFunction(experimentalData, PPIGraph,
  extraArgs)
}
\arguments{
\item{extraArgs}{}
}
\description{
Title
}

