% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{runTestOnData}
\alias{runTestOnData}
\title{Title}
\usage{
runTestOnData(data, normalSampleIndexes, diseaseSampleIndexes, testFunction,
  testFunctionExtraArgs)
}
\arguments{
\item{testFunctionExtraArgs}{}
}
\description{
Title
}

