% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{createIGraphObject}
\alias{createIGraphObject}
\title{Create graph object from gene interaction list}
\usage{
createIGraphObject(geneInteractionList)
}
\arguments{
\item{geneInteractionList}{- a data frame that contains the gene interaction
graph in two columns, the first being
the source HGNCSymbol and the second being the
targer HGNCSymbol of an. interaction}
}
\value{
A list containing igraph_object - the igraph object and
                          vertex_map - a frame containing node attributes
                          currently indegree and outdegree
}
\description{
Convert a data frame containing the gene interaction list into
a form usable by the other functions in the library. This is just
a wrapper around the igraph graph.edgelist function, which creates
and returns an igraph object with a data frame containing the in and
outdegrees.
}
\examples{
\dontrun{
     mitab <- downloadReactomeInteractions(data_folder="/tmp/")
     geneInteractionList <- getGeneInteractionListFromReactomeMITAB(mitab)
     ppiGraph<-createIGraphObject(geneInteractionList)
}

}
\references{
\url{http://igraph.org/r/doc/graph_from_edgelist.html}
}
\seealso{
\code{\link{getGeneInteractionListFromReactomeMITAB}}
}

