% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{KLDivergenceDistanceFunction}
\alias{KLDivergenceDistanceFunction}
\title{Title}
\usage{
KLDivergenceDistanceFunction(normalPageRanks, diseasePageRanks, extraArgs)
}
\arguments{
\item{extraArgs}{}
}
\description{
Title
}

