% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{meanAbsoluteDeviationDistanceFunction}
\alias{meanAbsoluteDeviationDistanceFunction}
\title{Title}
\usage{
meanAbsoluteDeviationDistanceFunction(normalPageRanks, diseasePageRanks,
  extraArgs)
}
\arguments{
\item{extraArgs}{}
}
\description{
Title
}

