% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllFunctions.R
\name{outdegreeNormalizedFCScoreFunction}
\alias{outdegreeNormalizedFCScoreFunction}
\title{Title}
\usage{
outdegreeNormalizedFCScoreFunction(experimentalData, PPIGraph, extraArgs)
}
\arguments{
\item{extraArgs}{}
}
\description{
Title
}

